refactor: break up ProcState
