+++ cli/tools/info.rs
@@ -36,9 +36,10 @@ pub async fn info(flags: Flags, info_flags: InfoFlags) -> Result<(), AnyError> {
   let ps = ProcState::from_flags(flags).await?;
   if let Some(specifier) = info_flags.file {
     let specifier = resolve_url_or_path(&specifier, ps.options.initial_cwd())?;
-    let mut loader = ps.create_graph_loader();
+    let mut loader = ps.module_graph_builder.create_graph_loader();
     loader.enable_loading_cache_info(); // for displaying the cache information
     let graph = ps
+      .module_graph_builder
       .create_graph_with_loader(vec![specifier], &mut loader)
       .await?;
